pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "playwrite-tutorial"
version = "0.1.0"
description = "Interactive Firefox REPL examples for learning Playwright with uv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.58.0",
    "pyyaml>=6.0",
]

[project.scripts]
ex01_repl = "playwrite_tutorial.ex01:main"
ex02_macro = "playwrite_tutorial.ex02:main"
ex03_tasks = "playwrite_tutorial.ex03:main"

[tool.hatch.build.targets.wheel]
packages = ["src/playwrite_tutorial"]
